requirements.txt
azure-core==1.30.2
azure-identity==1.17.1
azure-storage-blob==12.22.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
intel-openmp==2021.4.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.4.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
pywin32==306
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
six==1.16.0
sympy==1.13.1
tbb==2021.13.1
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.4.0
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3